refactor: test package 구성 변경 및 Car model test 추가
